we need two segment indices for periodic boundaries.
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/alugrid/trunk@1360 0d966ed9-3843-0410-af09-ebfb50bd7c74

+++ src/serial/gitter_sti.h
@@ -602,7 +602,7 @@ public :
     virtual int nEdges() const = 0;
     
     // return index of boundary segment 
-    virtual int segmentIndex () const { return -1; }
+    virtual int segmentIndex (const int) const { return -1; }
         
     // mark element for using iso8 rule 
     virtual int tagForGlobalRefinement () = 0 ;
